feat: extract GtinService to feedservice module
