+++ services/feedservice/README.md
@@ -1,0 +1,17 @@
+# Feed Service
+
+This module exposes utilities for generating EAN13 barcodes for GTINs.
+It can be reused by other microservices such as `ui` and `static`.
+
+## Features
+
+- Generate GTIN-13 barcode images using Barcode4J.
+
+## Build & Test
+
+```bash
+mvn clean install
+mvn test
+```
+
+See the [main project](../../README.md) for more information.
