+++ static/pom.xml
@@ -74,6 +74,11 @@
       <artifactId>serialisation</artifactId>
       <version>${global.version}</version>
     </dependency>
+    <dependency>
+      <groupId>org.open4goods</groupId>
+      <artifactId>feedservice</artifactId>
+      <version>${global.version}</version>
+    </dependency>
 
     <dependency>
       <groupId>org.springframework.boot</groupId>
